apps/Api.CardanoSounds/requirements.txt
arrow==1.1.1
arweave-python-client==1.0.14
asn1crypto==0.24.0
attrs==21.2.0
Automat==0.6.0
azure-core==1.16.0
azure-cosmos==4.2.0
blinker==1.4
bottle==0.12.19
certifi==2021.5.30
cffi==1.14.6
chardet==3.0.4
charset-normalizer==2.0.3
click==8.0.1
cloud-init==21.2
colorama==0.4.4
command-not-found==0.3
conan==1.39.0
configobj==5.0.6
constantly==15.1.0
cryptography==2.1.4
Cython==0.29.24
deprecation==2.0.7
dill==0.3.4
distro==1.6.0
distro-info===0.18ubuntu0.18.04.1
dlx==1.0.4
docplex==2.21.207
ecdsa==0.17.0
fastdtw==0.3.4
fasteners==0.16.3
fastjsonschema==2.15.1
Flask==2.0.1
future==0.18.2
h5py==3.2.1
httplib2==0.9.2
hyperlink==17.3.1
idna==3.2
incremental==16.10.1
inflection==0.5.1
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
jsonpatch==1.16
jsonpickle==2.0.0
jsonpointer==1.10
jsons==1.5.0
jsonschema==3.2.0
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
lxml==4.6.3
MarkupSafe==2.0.1
more-itertools==8.8.0
mpmath==1.2.1
multitasking==0.0.9
netifaces==0.10.4
node-semver==0.6.1
ntlm-auth==1.5.0
numpy==1.21.1
oauthlib==2.0.6
packaging==21.0
PAM==0.4.2
pandas==1.3.1
patch-ng==1.17.4
pexpect==4.2.1
pluginbase==1.0.1
ply==3.11
psutil==5.8.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pybind11==2.7.0
pycparser==2.20
pycrypto==2.6.1
pycryptodome==3.10.1
pydub==0.25.1
Pygments==2.9.0
PyGObject==3.26.1
pyinotify==0.9.6
PyJWT==1.7.1
PyNaCl==1.4.0
pyOpenSSL==17.5.0
pyparsing==2.4.7
pyrsistent==0.18.0
pyserial==3.4
python-apt==1.6.5+ubuntu0.6
python-constraint==1.4.0
python-dateutil==2.8.2
python-debian==0.1.32
python-jose==3.3.0
pytz==2021.1
pyxdg==0.25
PyYAML==5.4.1
qiskit==0.28.0
qiskit-aer==0.8.2
qiskit-aqua==0.9.4
qiskit-ibmq-provider==0.15.0
qiskit-ignis==0.6.0
qiskit-terra==0.18.0
qrandom==1.0
Quandl==3.6.1
redis==3.5.3
requests==2.26.0
requests-ntlm==1.1.0
requests-unixsocket==0.1.5
retworkx==0.9.0
rq==1.9.0
rsa==4.7.2
scikit-build==0.11.1
scikit-learn==0.24.2
scipy==1.7.0
SecretStorage==2.3.1
semantic-version==2.8.5
service-identity==16.0.0
setuptools-rust==0.12.1
simplejson==3.13.2
six==1.16.0
sos==4.1
ssh-import-id==5.7
symengine==0.7.2
sympy==1.8
systemd-python==234
threadpoolctl==2.2.0
toml==0.10.2
tqdm==4.61.2
tweedledum==1.1.0
Twisted==17.9.0
typish==1.9.2
ubuntu-advantage-tools==27.2
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.11
WALinuxAgent==2.2.45
websocket-client==1.1.0
Werkzeug==2.0.1
yfinance==0.1.62
zope.interface==4.3.2
